requirements.txt
requests==2.31.0
psutil==5.9.8
pywin32==311
rapidocr_onnxruntime==1.4.4
PyQt5==5.15.10
Pillow==12.1.1
pyautogui==0.9.54